:sparkles: Introduced DesignsystemetKomponentIntro and DesignsystemetThumbnail components for enhanced component documentation
